changes to cottey scope search form, and all scopes Numbers menu behavior
There was discussion about the Numbers menu being expanded by default and the addition of some indicator that it could be collapsed or re-expanded. plus and minus signs, some css, and changes to the menu-collapse.js file were made. The menu-copplapse.js file shoudl go away and the functions be moved to a more generic script.